+++ Changes.md
@@ -4,8 +4,12 @@
 Improvements
 ------------
 
+- RenderMan : Added light linking support.
+- Cycles : Added light linking support.
 - VisualiserTool : Added new visualisation for orientation (Quatf) data.
 - PrimitiveInspector : Changed column order for quaternions to match Imath's conventions.
+- AttributeEditor : Improved widgets used for editing `linkedLights` and `filteredLights` attributes.
+- Viewer : Improved light linking performance.
 
 Fixes
 -----